# cmd/campaign-service/kodata/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: LFX V2 - Campaign Service
    description: A collection of service endpoints to support Marketing Operations campaign creation and management.
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /projects/{project_id}/briefs/{brief_id}/campaigns/{campaign_id}:
        get:
            tags:
                - lfx-v2-campaign-service-briefs
            summary: get-campaign lfx-v2-campaign-service-briefs
            description: Get one campaign under a brief; returns ETag.
            operationId: lfx-v2-campaign-service-briefs#get-campaign
            parameters:
                - name: project_id
                  in: path
                  description: Project UUID or slug that scopes the connection
                  required: true
                  type: string
                - name: brief_id
                  in: path
                  description: Brief UUID
                  required: true
                  type: string
                  format: uuid
                - name: campaign_id
                  in: path
                  description: Campaign UUID
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Campaign'
                        required:
                            - id
                            - project_id
                            - brief_id
                            - platform
                            - campaign_name
                            - status
                            - version
                    headers:
                        ETag:
                            description: ETag header value (mirrors version)
                            type: string
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ConnServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        put:
            tags:
                - lfx-v2-campaign-service-briefs
            summary: update-campaign lfx-v2-campaign-service-briefs
            description: Replace a campaign (requires If-Match).
            operationId: lfx-v2-campaign-service-briefs#update-campaign
            parameters:
                - name: project_id
                  in: path
                  description: Project UUID or slug that scopes the connection
                  required: true
                  type: string
                - name: brief_id
                  in: path
                  description: Brief UUID
                  required: true
                  type: string
                  format: uuid
                - name: campaign_id
                  in: path
                  description: Campaign UUID
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
                - name: If-Match
                  in: header
                  description: If-Match header carrying the current ETag/version
                  required: false
                  type: string
                - name: Update-CampaignRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/LfxV2CampaignServiceBriefsUpdateCampaignRequestBody'
                    required:
                        - campaign
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Campaign'
                        required:
                            - id
                            - project_id
                            - brief_id
                            - platform
                            - campaign_name
                            - status
                            - version
                    headers:
                        ETag:
                            description: ETag header value (mirrors version)
                            type: string
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - code
                            - message
                "412":
                    description: Precondition Failed response.
                    schema:
                        $ref: '#/definitions/PreconditionFailedError'
                        required:
                            - code
                            - message
                "428":
                    description: Precondition Required response.
                    schema:
                        $ref: '#/definitions/PreconditionRequiredError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ConnServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
definitions:
...
    BadRequestError:
        title: BadRequestError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "400"
            message:
                type: string
                description: Error message
                example: The request was invalid.
        description: Bad request
        example:
            code: "400"
            message: The request was invalid.
        required:
            - code
            - message
...
    Campaign:
        title: Campaign
        type: object
        properties:
            brief_id:
                type: string
                description: Parent brief
                example: Ut architecto molestiae cum.
            campaign_name:
                type: string
                description: Campaign name
                example: Quo veritatis nisi velit distinctio animi.
            id:
                type: string
                description: Campaign UUID
                example: Id sint suscipit.
            platform:
                type: string
                description: Channel
                example: Perspiciatis consequatur beatae sunt officiis sequi qui.
            platform_campaign_id:
                type: string
                description: ID returned by the ad platform
                example: Sapiente perspiciatis dolores dignissimos soluta non vel.
            project_id:
                type: string
                description: Owning project
                example: Ipsa voluptas temporibus non neque.
            status:
                type: string
                description: Campaign status
                example: Provident rerum et voluptas tenetur esse.
            version:
                type: integer
                description: Optimistic-concurrency version
                example: 4427900281484301982
                format: int64
        example:
            brief_id: Eveniet aliquid quia aut impedit a veritatis.
            campaign_name: Non illo ab in voluptatem minus est.
            id: Sed minima nihil voluptatem quaerat ipsa voluptatem.
            platform: Consequuntur sed ea sunt debitis.
            platform_campaign_id: Voluptas qui voluptas veniam sint voluptas.
            project_id: Dolor molestiae quia voluptatem deserunt libero.
            status: Molestiae soluta vel ad et in voluptatem.
            version: 481010994209444199
        required:
            - id
            - project_id
            - brief_id
            - platform
            - campaign_name
            - status
            - version
...
    CampaignUpdateInput:
        title: CampaignUpdateInput
        type: object
        properties:
            campaign_name:
                type: string
                description: Campaign name
                example: Ea dolor cumque dolores qui.
            config:
                description: Campaign configuration snapshot
                example: Aspernatur culpa sed non qui ad.
            status:
                type: string
                description: Campaign status
                example: Ratione dicta.
        example:
            campaign_name: Voluptas qui.
            config: Possimus totam occaecati enim eaque.
            status: Rerum voluptatem saepe eligendi eum non.
        required:
            - campaign_name
            - status
    ConflictError:
        title: ConflictError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "409"
            message:
                type: string
                description: Error message
                example: A connection for this provider already exists on the project.
        description: Conflict
        example:
            code: "409"
            message: A connection for this provider already exists on the project.
        required:
            - code
            - message
    ConnServiceUnavailableError:
        title: ConnServiceUnavailableError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "503"
            message:
                type: string
                description: Error message
                example: The service is unavailable.
        description: Service unavailable
        example:
            code: "503"
            message: The service is unavailable.
        required:
            - code
            - message
...
    InternalServerError:
        title: InternalServerError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "500"
            message:
                type: string
                description: Error message
                example: An internal server error occurred.
        description: Internal server error
        example:
            code: "500"
            message: An internal server error occurred.
        required:
            - code
            - message
...
    LfxV2CampaignServiceBriefsUpdateCampaignRequestBody:
        title: LfxV2CampaignServiceBriefsUpdateCampaignRequestBody
        type: object
        properties:
            campaign:
                $ref: '#/definitions/CampaignUpdateInput'
        example:
            campaign:
                campaign_name: Deserunt vero ducimus distinctio deserunt.
                config: Sed et.
                status: Magni aut nisi est sit sint.
        required:
            - campaign
...
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "404"
            message:
                type: string
                description: Error message
                example: The connection was not found.
        description: Resource not found
        example:
            code: "404"
            message: The connection was not found.
        required:
            - code
            - message
...
    PreconditionFailedError:
        title: PreconditionFailedError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "412"
            message:
                type: string
                description: Error message
                example: The supplied ETag does not match the current version.
        description: ETag mismatch
        example:
            code: "412"
            message: The supplied ETag does not match the current version.
        required:
            - code
            - message
    PreconditionRequiredError:
        title: PreconditionRequiredError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: "428"
            message:
                type: string
                description: Error message
                example: An If-Match header is required.
        description: If-Match header required
        example:
            code: "428"
            message: An If-Match header is required.
        required:
            - code
            - message
...
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: JWT issued by Heimdall; audience is this service.
        name: Authorization
        in: header