# gen/http/lfx_v2_campaign_service_metrics/server/encode_decode.go
// Code generated by goa v3.25.3, DO NOT EDIT.
//
// lfx-v2-campaign-service-metrics HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/linuxfoundation/lfx-v2-campaign-service/design

package server

import (
	"context"
	"errors"
	"net/http"
	"strings"

	lfxv2campaignservicemetrics "github.com/linuxfoundation/lfx-v2-campaign-service/gen/lfx_v2_campaign_service_metrics"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetCampaignMetricsResponse returns an encoder for responses returned
// by the lfx-v2-campaign-service-metrics get-campaign-metrics endpoint.
func EncodeGetCampaignMetricsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*lfxv2campaignservicemetrics.GetCampaignMetricsResult)
		enc := encoder(ctx, w)
		body := NewGetCampaignMetricsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetCampaignMetricsRequest returns a decoder for requests sent to the
// lfx-v2-campaign-service-metrics get-campaign-metrics endpoint.
func DecodeGetCampaignMetricsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*lfxv2campaignservicemetrics.GetCampaignMetricsPayload, error) {
	return func(r *http.Request) (*lfxv2campaignservicemetrics.GetCampaignMetricsPayload, error) {
		var payload *lfxv2campaignservicemetrics.GetCampaignMetricsPayload
		var (
			projectID   string
			briefID     string
			campaignID  string
			from        *string
			to          *string
			bearerToken *string
			err         error

			params = mux.Vars(r)
		)
		projectID = params["project_id"]
		briefID = params["brief_id"]
		err = goa.MergeErrors(err, goa.ValidateFormat("brief_id", briefID, goa.FormatUUID))
		campaignID = params["campaign_id"]
		err = goa.MergeErrors(err, goa.ValidateFormat("campaign_id", campaignID, goa.FormatUUID))
		qp := r.URL.Query()
		fromRaw := qp.Get("from")
		if fromRaw != "" {
			from = &fromRaw
		}
		if from != nil {
			err = goa.MergeErrors(err, goa.ValidatePattern("from", *from, "^\\d{4}-\\d{2}-\\d{2}$"))
		}
		toRaw := qp.Get("to")
		if toRaw != "" {
			to = &toRaw
		}
		if to != nil {
			err = goa.MergeErrors(err, goa.ValidatePattern("to", *to, "^\\d{4}-\\d{2}-\\d{2}$"))
		}
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		if err != nil {
			return payload, err
		}
		payload = NewGetCampaignMetricsPayload(projectID, briefID, campaignID, from, to, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeGetCampaignMetricsError returns an encoder for errors returned by the
// get-campaign-metrics lfx-v2-campaign-service-metrics endpoint.
func EncodeGetCampaignMetricsError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "BadRequest":
			var res *lfxv2campaignservicemetrics.BadRequestError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetCampaignMetricsBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "Conflict":
			var res *lfxv2campaignservicemetrics.ConflictError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetCampaignMetricsConflictResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return enc.Encode(body)
		case "ServiceUnavailable":
			var res *lfxv2campaignservicemetrics.ConnServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetCampaignMetricsServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		case "InternalServerError":
			var res *lfxv2campaignservicemetrics.InternalServerError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetCampaignMetricsInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "NotFound":
			var res *lfxv2campaignservicemetrics.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetCampaignMetricsNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalLfxv2campaignservicemetricsCampaignMetricToCampaignMetricResponseBody
// builds a value of type *CampaignMetricResponseBody from a value of type
// *lfxv2campaignservicemetrics.CampaignMetric.
func marshalLfxv2campaignservicemetricsCampaignMetricToCampaignMetricResponseBody(v *lfxv2campaignservicemetrics.CampaignMetric) *CampaignMetricResponseBody {
	res := &CampaignMetricResponseBody{
		MetricDate:       v.MetricDate,
		Platform:         v.Platform,
		Impressions:      v.Impressions,
		Clicks:           v.Clicks,
		Spend:            v.Spend,
		Conversions:      v.Conversions,
		Currency:         v.Currency,
		AttributionBasis: v.AttributionBasis,
		FetchedAt:        v.FetchedAt,
	}

	return res
}

// marshalLfxv2campaignservicemetricsMetricsSummaryToMetricsSummaryResponseBody
// builds a value of type *MetricsSummaryResponseBody from a value of type
// *lfxv2campaignservicemetrics.MetricsSummary.
func marshalLfxv2campaignservicemetricsMetricsSummaryToMetricsSummaryResponseBody(v *lfxv2campaignservicemetrics.MetricsSummary) *MetricsSummaryResponseBody {
	res := &MetricsSummaryResponseBody{
		Impressions:           v.Impressions,
		Clicks:                v.Clicks,
		Spend:                 v.Spend,
		Currency:              v.Currency,
		CurrencyUniform:       v.CurrencyUniform,
		Conversions:           v.Conversions,
		AttributionBasis:      v.AttributionBasis,
		ConversionsComparable: v.ConversionsComparable,
		RowCount:              v.RowCount,
	}

	return res
}